engine: '^g**' axis-fill seeds + fractional rule weights
'*' in a seed position fills that axis (both = flood the field) - the
dual of the bare '^' clear, replacing the 4-rule background-cascade
idiom. Weights parse as decimals (scores stay integer), so rare events
(lightning 0.01) no longer inflate every other weight; selection math
was already double, integer-weight programs replay bit-identically
(verified against recorded lambda_c data). forest2 v3: 11 -> 7 rules.
Docs: GRAMMAR.v2.md + GRAMMAR.md.

+++ src/grammar.h
@@ -83,7 +83,7 @@ public:
         wchar_t ctx;
         wchar_t rep;
         wchar_t ctxrep;
-        int weight;
+        double weight;
         wchar_t sound;
         bool load;
         bool engine_action;
@@ -175,7 +175,7 @@ struct RuleApplication {
     std::pair<int, int> position;
     Grammar2D::Rule rule;
     size_t rule_index;
-    int weight;
+    double weight;
 };
 
 // Per-cell match outcome reported during an "explain mode" dry-run
